requirements.txt
Flask==3.0.0
gunicorn==21.2.0
fyers-apiv3==3.1.7
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0